refactor(github): move github mcp module into package
