+++ integrations/_catalog_impl.py
@@ -50,8 +50,8 @@ from integrations.dagster import classify as _classify_dagster
 from integrations.datadog import classify as _classify_datadog
 from integrations.discord import classify as _classify_discord
 from integrations.effective_models import EffectiveIntegrations
-from integrations.github_mcp import build_github_mcp_config
-from integrations.github_mcp import classify as _classify_github
+from integrations.github.mcp import build_github_mcp_config
+from integrations.github.mcp import classify as _classify_github
 from integrations.gitlab import DEFAULT_GITLAB_BASE_URL, build_gitlab_config
 from integrations.gitlab import classify as _classify_gitlab
 from integrations.grafana import classify as _classify_grafana
