+++ integrations/openclaw/__init__.py
@@ -420,7 +420,7 @@ async def _open_openclaw_session(config: OpenClawConfig) -> AsyncIterator[Client
                 env={
                     **os.environ,
                     # Suppress terminal control codes so the MCP server's stdout
-                    # stays clean JSON-RPC (mirrors github_mcp.py mitigation).
+                    # stays clean JSON-RPC (mirrors integrations/github/mcp.py mitigation).
                     "NO_COLOR": "1",
                     "TERM": "dumb",
                     **({"OPENCLAW_AUTH_TOKEN": config.auth_token} if config.auth_token else {}),
