+++ integrations/sentry_mcp/__init__.py
@@ -259,7 +259,7 @@ async def _open_sentry_mcp_session(config: SentryMCPConfig) -> AsyncIterator[Cli
                 env={
                     **os.environ,
                     # Suppress terminal control codes so the MCP server's stdout
-                    # stays clean JSON-RPC (mirrors github_mcp.py mitigation).
+                    # stays clean JSON-RPC (mirrors integrations/github/mcp.py mitigation).
                     "NO_COLOR": "1",
                     "TERM": "dumb",
                     **({"SENTRY_ACCESS_TOKEN": config.auth_token} if config.auth_token else {}),
