+++ surfaces/cli/constants.py
@@ -44,7 +44,7 @@ def __getattr__(name: str) -> tuple[str, ...]:
     # sourced from the runtime integration registry so the CLI's positional-arg
     # `click.Choice` validators stay in sync with what cmd_setup / cmd_verify
     # can actually dispatch. Eagerly importing `integrations.registry` here
-    # creates a circular import (registry -> verifiers -> github_mcp ->
+    # creates a circular import (registry -> verifiers -> integrations.github.mcp ->
     # cli.*). Deferring to first access lets `cli` finish
     # bootstrapping. See #1973 (verify) and #2537 (setup).
     if name == "SETUP_SERVICES":
